+++ ps2000a/connection.go
@@ -96,10 +96,11 @@ func getValuesBulk(m *genericps.GetValuesBulkMsg) {
 		err          error
 		numOfSamples uint32
 	)
-	numOfSamples, err = ps2000aGetValuesBulk(m.Handle(), m.ReqNumOfSamples, m.FromSegmentIndex, m.ToSegmentIndex,
-		m.DownSampleRatio, RatioMode(m.DownSampleRatioMode), m.Overflow)
+	numOfSamples, err = ps2000aGetValuesBulk(m.Handle(), uint32(m.ReqNumOfSamples),
+		uint32(m.FromSegmentIndex), uint32(m.ToSegmentIndex),
+		uint32(m.DownSampleRatio), RatioMode(m.DownSampleRatioMode), m.Overflow)
 	response := m.Rsp().(*genericps.GetValuesBulkRsp)
-	response.NumOfSamples = numOfSamples
+	response.NumOfSamples = uint64(numOfSamples)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
 }
@@ -180,7 +181,7 @@ func getMaxSegments(m *genericps.GetMaxSegmentsMsg) {
 func getNumberOfCaptures(m *genericps.GetNumOfCapturesMsg) {
 	numOfCaptures, err := ps2000aGetNumOfCaptures(m.Handle())
 	response := m.Rsp().(*genericps.GetNumOfCapturesRsp)
-	response.NCaptures = numOfCaptures
+	response.NCaptures = uint64(numOfCaptures)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
 }
@@ -188,7 +189,7 @@ func getNumberOfCaptures(m *genericps.GetNumOfCapturesMsg) {
 func getNumberOfProcessedCaptures(m *genericps.GetNumOfProcessedCapturesMsg) {
 	numOfCaptures, err := ps2000aGetNumOfProcessedCaptures(m.Handle())
 	response := m.Rsp().(*genericps.GetNumOfProcessedCapturesRsp)
-	response.NCaptures = numOfCaptures
+	response.NCaptures = uint64(numOfCaptures)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
 }
@@ -204,19 +205,21 @@ func getStreamingLatestValues(m *genericps.GetStreamingLatestValuesMsg) {
 }
 
 func getTimebase(m *genericps.GetTimebaseMsg) {
-	timeIntervalNanoseconds, maxSamples, err := ps2000aGetTimebase(m.Handle(), m.TimeBase, m.NumOfSamples, m.OverSample, m.SegmentIndex)
+	timeIntervalNanoseconds, maxSamples, err := ps2000aGetTimebase(m.Handle(),
+		uint32(m.TimeBase), uint32(m.NumOfSamples), m.OverSample, uint32(m.SegmentIndex))
 	response := m.Rsp().(*genericps.GetTimebaseRsp)
-	response.TimeIntervalNanoseconds = timeIntervalNanoseconds
-	response.MaxSamples = maxSamples
+	response.TimeIntervalNanoseconds = float64(timeIntervalNanoseconds)
+	response.MaxSamples = uint64(maxSamples)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
 }
 
 func getTimebase2(m *genericps.GetTimebase2Msg) {
-	timeIntervalNanoseconds, maxSamples, err := ps2000aGetTimebase2(m.Handle(), m.TimeBase, m.NumOfSamples, m.OverSample, m.SegmentIndex)
+	timeIntervalNanoseconds, maxSamples, err := ps2000aGetTimebase2(m.Handle(),
+		uint32(m.TimeBase), int32(m.NumOfSamples), m.OverSample, uint32(m.SegmentIndex))
 	response := m.Rsp().(*genericps.GetTimebase2Rsp)
-	response.TimeIntervalNanoseconds = timeIntervalNanoseconds
-	response.MaxSamples = maxSamples
+	response.TimeIntervalNanoseconds = float64(timeIntervalNanoseconds)
+	response.MaxSamples = uint64(maxSamples)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
 }
@@ -252,7 +255,8 @@ func setDataBuffer(m *genericps.SetDataBufferMsg) {
 	var (
 		err error
 	)
-	err = ps2000aSetDataBuffer(m.Handle(), ChannelId(m.Ch), m.BufferIn, m.SegmentIndex, RatioMode(m.Mode))
+	err = ps2000aSetDataBuffer(m.Handle(), ChannelId(m.Ch), m.BufferIn,
+		uint32(m.SegmentIndex), RatioMode(m.Mode))
 	response := m.Rsp().(*genericps.SetDataBufferRsp)
 	response.SetStatus(err)
 	m.RspCh() <- struct{}{}
@@ -322,8 +326,10 @@ func runBlock(m *genericps.RunBlockMsg) {
 		err              error
 		timeIndisposedMs int32
 	)
-	timeIndisposedMs, err = ps2000aRunBlock(m.Handle(), m.NumOfPreTriggerSamples, m.NumOfPostTriggerSamples,
-		m.TimeBase, m.OverSample, m.SegmentIndex, BlockReady(m.LpBlockReadyGoPar), m.Param)
+	timeIndisposedMs, err = ps2000aRunBlock(m.Handle(), int32(m.NumOfPreTriggerSamples),
+		int32(m.NumOfPostTriggerSamples),
+		uint32(m.TimeBase), m.OverSample, uint32(m.SegmentIndex),
+		BlockReady(m.LpBlockReadyGoPar), m.Param)
 	response := m.Rsp().(*genericps.RunBlockRsp)
 	response.TimeIndisposedMs = timeIndisposedMs
 	response.SetStatus(err)
@@ -481,7 +487,7 @@ func getTriggerTimeOffset(m *genericps.GetTriggerTimeOffsetMsg) {
 		timeUpper, timeLower uint32
 		timeUnits            TimeUnits
 	)
-	timeUpper, timeLower, timeUnits, err = ps2000aGetTriggerTimeOffset(m.Handle(), m.SegmentIndex)
+	timeUpper, timeLower, timeUnits, err = ps2000aGetTriggerTimeOffset(m.Handle(), uint32(m.SegmentIndex))
 	response := m.Rsp().(*genericps.GetTriggerTimeOffsetRsp)
 	response.TimeLower = timeLower
 	response.TimeUnits = genericps.TimeUnits(timeUnits)
@@ -555,7 +561,7 @@ func memorySegments(m *genericps.MemorySegmentsMsg) {
 	nMaxSamples, err := ps2000aMemorySegments(m.Handle(), m.NSegments)
 	response := m.Rsp().(*genericps.MemorySegmentsRsp)
 	response.SetStatus(err)
-	response.NMaxSamples = nMaxSamples
+	response.NMaxSamples = uint64(nMaxSamples)
 	m.RspCh() <- struct{}{}
 }
